w.policy

+++ app/Http/Controllers/WorkerController.php
@@ -61,11 +61,15 @@ class WorkerController extends Controller
 
     public function create()
     {
+        $this->authorize('create',Worker::class);
         return view('worker.create');
     }
 
     public function store(StoreRequest $request)
     {
+
+        $this->authorize('create',Worker::class);
+
         $data = $request->validated();
         $data['is_married'] = isset($data['is_married']);
 
@@ -76,11 +80,14 @@ class WorkerController extends Controller
 
     public function edit(Worker $worker)
     {
+        $this->authorize('update', $worker);
         return view('worker.edit',compact('worker'));
     }
 
     public function update(UpdateRequest $request, Worker $worker)
     {
+        $this->authorize('update', $worker);
+
         $data = $request->validated();
         $data['is_married'] = isset($data['is_married']);
         $worker ->update($data);
@@ -90,6 +97,7 @@ class WorkerController extends Controller
     }
     public function destroy(Worker $worker)
     {
+        $this->authorize('delete', $worker);
         $worker->delete();
         return redirect()->route('workers.index');
     }
